+++ examples/advanced/procedural/voronoi.py
@@ -5,7 +5,7 @@ from window import Window
 window = Window((720, 720))
 ctx = zengl.context()
 
-image = ctx.image(window.size, 'rgba8unorm-srgb')
+image = ctx.image(window.size, 'rgba8unorm')
 
 pipeline = ctx.pipeline(
     vertex_shader='''
@@ -89,7 +89,7 @@ pipeline = ctx.pipeline(
             mt.randomn = 1.0;
 
             vec3 color = voronoi_color(uv, mt);
-            out_color = vec4(color, 1.0);
+            out_color = vec4(pow(color, vec3(1.0 / 2.2)), 1.0);
         }
     ''',
     uniforms={