+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 # [main](https://github.com/szabolcsdombi/zengl/compare/2.2.2...main)
 
+- Removed srgb image formats
+
 # [2.2.2](https://github.com/szabolcsdombi/zengl/compare/2.2.1...2.2.2)
 
 - Improved typehints and documentation